clean up main
